% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{align_columns}
\alias{align_columns}
\title{Align Estimated and True Matrices Using the Hungarian Algorithm}
\usage{
align_columns(D_est, D_true)
}
\arguments{
\item{D_est}{A numeric matrix of estimated components (n x k).}

\item{D_true}{A numeric matrix of true components (n x k).}
}
\value{
A numeric matrix with columns of \code{D_est} aligned to \code{D_true}.
}
\description{
This function aligns the columns of an estimated matrix with those of the true matrix by maximizing the absolute correlation between columns using the Hungarian algorithm.
}
\examples{
# Assuming D_est and D_true are matrices
D_aligned <- align_columns(D_est, D_true)
}
