% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{avgOICAruns}
\alias{avgOICAruns}
\title{Average Multiple OICA Runs to Obtain a Consensus Mixing Matrix}
\usage{
avgOICAruns(result, num_runs, p, k, maxit = 2000)
}
\arguments{
\item{result}{A list object returned by the \code{overica} function with multiple runs, containing the estimated mixing matrices.}

\item{num_runs}{An integer specifying the number of runs.}

\item{p}{An integer specifying the number of observed variables.}

\item{k}{An integer specifying the number of latent variables.}

\item{maxit}{Maximum number of iterations for the constrained clustering algorithm}
}
\value{
A matrix representing the averaged mixing matrix after aligning and averaging over runs.
}
\description{
This function averages the estimated mixing matrices from multiple runs of the \code{overica} function.
It aligns components across runs, accounting for sign ambiguity inherent in ICA, by clustering the
components and combining sign-flipped versions to obtain a consensus mixing matrix.
}
\examples{
\dontrun{
# Assuming 'result' is the output from overica with multiple runs
p <- ncol(data)
k <- 5
num_runs <- 10
averaged_A <- avgOICAruns(result, num_runs, p, k)
}
}
