% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_BIC}
\alias{compute_BIC}
\title{Compute pseudo BIC for an OLS-style reconstruction loss}
\usage{
compute_BIC(raw_loss, A_est, B_est, k, threshold = 0.001, third = TRUE)
}
\arguments{
\item{raw_loss}{A numeric value representing the mean squared error (MSE) of the reconstruction.}

\item{A_est}{A torch tensor (or array) for the estimated matrix A.}

\item{B_est}{A torch tensor (or array) for the estimated matrix B.}

\item{threshold}{Numeric; entries smaller than this in absolute value are treated as zero.}

\item{third}{Are third moments to be considered?}

\item{n_used}{Number of observations used in the loss calculation.}
}
\value{
A list with elements:
\item{BIC}{The computed BIC value.}
\item{SSE}{The sum of squared errors.}
\item{k_eff}{The effective number of parameters.}
}
\description{
Given the mean-squared reconstruction error (raw_loss), the estimated matrices A and B,
and the number of observations (n_used), this function computes the sum of squared errors (SSE)
as SSE = raw_loss * n_used and counts the effective number of parameters (k_eff) as the number of
entries in A and B whose absolute value exceeds a given threshold.
Then the BIC is computed as:
}
\details{
BIC = n_used * log(SSE/n_used) + log(n_used) * k_eff
}
