% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_matrix}
\alias{generate_matrix}
\title{Generate a Matrix of Non-Gaussian Variables}
\usage{
generate_matrix(n, k)
}
\arguments{
\item{n}{An integer specifying the number of observations (rows).}

\item{k}{An integer specifying the number of variables (columns).}
}
\value{
A numeric matrix of size \code{n x k} with non-Gaussian variables.
}
\description{
This function generates a matrix where each column is a non-Gaussian random variable of a specified type.
}
\examples{
z <- generate_matrix(1000, 5)
}
