% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_non_gaussian}
\alias{generate_non_gaussian}
\title{Generate Non-Gaussian Random Variables}
\usage{
generate_non_gaussian(
  n,
  type = c("skew_positive", "skew_negative", "kurtotic")
)
}
\arguments{
\item{n}{An integer specifying the number of observations to generate.}

\item{type}{A character string specifying the type of non-Gaussian distribution. One of "skew_positive", "skew_negative", or "kurtotic".}
}
\value{
A numeric vector of length \code{n} with non-Gaussian random variables.
}
\description{
This function generates non-Gaussian random variables of specified type.
}
\examples{
x <- generate_non_gaussian(1000, type = "skew_positive")
}
