% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{torch_unique_4th_central_moments}
\alias{torch_unique_4th_central_moments}
\title{Compute Unique 4th-Order Central Moments (i <= j <= k <= l)}
\usage{
torch_unique_4th_central_moments(X)
}
\arguments{
\item{X}{A torch tensor of shape (n, p).}
}
\value{
A 1D torch tensor of length choose(p+3, 4).
}
\description{
For data X of shape (n, p), we first center each column, i.e. X_centered\link{,i} = X\link{,i} - mean_i.
Then for each 1 <= i <= j <= k <= l <= p, we compute
E\link{ (X_i - mu_i) (X_j - mu_j) (X_k - mu_k) (X_l - mu_l) }.
}
\details{
This yields a total of choose(p+3, 4) distinct moments (a 1D torch vector).
}
